% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{allNA}
\alias{allNA}
\title{Are all elements `NA`?}
\usage{
allNA(x)
}
\arguments{
\item{x}{data vector}
}
\value{
A logical variable indicating all elements are 'NA'
}
\description{
Return a logical indicating if all elements are `NA`
}
\examples{
allNA(c(NA, NA, 1, NA)) ## FALSE
}
