% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{any_finite}
\alias{any_finite}
\title{Rows with finite values}
\usage{
any_finite(Y)
}
\arguments{
\item{Y}{matrix like data object}
}
\value{
Indexes of rows with at least one finite value
}
\description{
Return indexes of rows with at least one finite value
}
\examples{
X <- matrix(1,10,2)
X[3,] <- NA
any_finite(X)
}
