% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{count_obs}
\alias{count_obs}
\title{Count observations}
\usage{
count_obs(x)
}
\arguments{
\item{x}{data vector}
}
\value{
The Number of observations
}
\description{
Return the number of finite observations in `x`
}
\examples{
count_obs(c(1,3,5,7,9,NA)) # 5
}
