% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateutils.R
\name{day}
\alias{day}
\title{Return the day of a Date value}
\usage{
day(date)
}
\arguments{
\item{date}{date value formated as.Date()}
}
\value{
the day of the date (integer)
}
\description{
Return the day of a Date value as an integer
}
\examples{
day(as.Date("2019-09-15")) ## 15
}
