% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Diff}
\alias{Diff}
\title{Difference data}
\usage{
Diff(x, lag = 1)
}
\arguments{
\item{x}{data}

\item{lag}{number of lags to use}
}
\value{
Differenced data
}
\description{
Wrapper for `diff()` maintaining the same number of observations in `x`
}
\examples{
Diff(c(100,50,100,20,100,110))
}
