% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateutils.R
\name{end_next_month}
\alias{end_next_month}
\title{End of next month date}
\usage{
end_next_month(date)
}
\arguments{
\item{date}{date value formated as.Date()}
}
\description{
Return the date of the last day of the following month
}
