% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateutils.R
\name{end_of_month}
\alias{end_of_month}
\title{End of  month date}
\usage{
end_of_month(date)
}
\arguments{
\item{date}{date value formated as.Date()}
}
\description{
Return the date of the last day of the month
}
