% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateutils.R
\name{end_of_quarter}
\alias{end_of_quarter}
\title{End of quarter date}
\usage{
end_of_quarter(date)
}
\arguments{
\item{date}{date value formated as.Date()}
}
\description{
Return the date of the last day of the quarter
}
