% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateutils.R
\name{end_of_year}
\alias{end_of_year}
\title{End of Year}
\usage{
end_of_year(dates)
}
\arguments{
\item{dates}{Transition matrix from a VAR model}
}
\description{
Find the end of year for a vector of dates
}
