% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateutils.R
\name{end_previous_month}
\alias{end_previous_month}
\title{End of previous month date}
\usage{
end_previous_month(date)
}
\arguments{
\item{date}{date value formated as.Date()}
}
\description{
Return the date of the last day of the previous month
}
