% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateutils.R
\name{end_previous_quarter}
\alias{end_previous_quarter}
\title{End of previous quarter date}
\usage{
end_previous_quarter(date)
}
\arguments{
\item{date}{date value formated as.Date()}
}
\description{
Return the date of the last day of the previous quarter
}
