% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_basic_character}
\alias{extract_basic_character}
\title{Extract characters}
\usage{
extract_basic_character(x)
}
\arguments{
\item{x}{object containing character (and other) values}
}
\description{
Extract character values from x excluding space and underscore
}
