% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_character}
\alias{extract_character}
\title{Extract character values}
\usage{
extract_character(x)
}
\arguments{
\item{x}{object containing character values}
}
\value{
Character valus from the object
}
\description{
Extract character values from x including space and underscore
}
\examples{
extract_character(c("this_1one", "abc123")) ## c("this_one", "abc")
}
