% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_numeric}
\alias{extract_numeric}
\title{Extract numeric values}
\usage{
extract_numeric(x)
}
\arguments{
\item{x}{object containing numeric (and other) values}
}
\value{
Numeric values from the object
}
\description{
Extract numeric values from x
}
\examples{
extract_numeric(c("7+5", "abc123")) ## c(75, 123)
}
