% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fill_daily_dates}
\alias{fill_daily_dates}
\title{Normalize daily dates in a data.table}
\usage{
fill_daily_dates(DT, date_var = "ref_date")
}
\arguments{
\item{DT}{a data.table}

\item{date_var}{name of date column}
}
\description{
Normalize daily dates in a data.table to prevent missing days
}
