% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateutils.R
\name{fill_forward}
\alias{fill_forward}
\title{Fill Forward}
\usage{
fill_forward(x)
}
\arguments{
\item{x}{Transition matrix from a VAR model}
}
\description{
Fill missing observations forward using the last finite observation
}
