% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{first_of_month}
\alias{first_of_month}
\title{First of month}
\usage{
first_of_month(dates)
}
\arguments{
\item{dates}{A sequence of dates in `as.Date()` format}
}
\description{
Return the first day of the month for each date in `dates`
}
