% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_weekly}
\alias{format_weekly}
\title{Format weekly frequency data}
\usage{
format_weekly(dt, date_name = "ref_date")
}
\arguments{
\item{dt}{wide format data.table}

\item{date_name}{name of column containing dates in `as.Date()` format}
}
\description{
Regularlize data to weekly frequency so that all observations fall at the end of the week
}
