% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateutils.R
\docType{data}
\name{fred}
\alias{fred}
\title{Sample mixed frequency data from FRED}
\description{
Sample mixed frequency data from FRED
}
\references{
\url{https://fred.stlouisfed.org/}
}
\author{
Seth Leonard \email{seth@macroeconomicdata.com}
}
\keyword{data}
