% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateutils.R
\docType{data}
\name{fredlib}
\alias{fredlib}
\title{Library of metadata for mixed frequency dataset `fred`}
\description{
Library of metadata for mixed frequency dataset `fred`
}
\references{
\url{https://fred.stlouisfed.org/}
}
\author{
Seth Leonard \email{seth@macroeconomicdata.com}
}
\keyword{data}
