% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_from_list}
\alias{get_from_list}
\title{Get from list}
\usage{
get_from_list(lst, what)
}
\arguments{
\item{lst}{list}

\item{what}{object to retrieve (by name or index)}
}
\description{
Retrieve object `what` from `lst`
}
