% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{index_by_friday}
\alias{index_by_friday}
\title{Find the Friday in a given week}
\usage{
index_by_friday(dates)
}
\arguments{
\item{dates}{vector of dates}
}
\value{
The date of the Friday in the week of the given date
}
\description{
Find the Friday in a given week from a sequence of Dates
Vectors should be in as.Date() format
}
\examples{
dates <- seq.Date(from = as.Date("2020-09-21"),
                  by = "week", length.out = 10)
fridays <- index_by_friday(dates)
weekdays(fridays)
}
