% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{last_in_quarter}
\alias{last_in_quarter}
\title{Last date in the week}
\usage{
last_in_quarter(dates)
}
\arguments{
\item{dates}{A sequence of dates in `as.Date()` format}
}
\description{
Return the latest date in the quarter fop the values in `dates`
}
