% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seas_utils.R
\name{match_ts_dates}
\alias{match_ts_dates}
\title{Match dates between two timeseries}
\usage{
match_ts_dates(old_ts, new_ts)
}
\arguments{
\item{old_ts}{timeseries data}

\item{new_ts}{timeseries data}
}
\description{
Find values in `new_ts` that correspond to dates in `old_ts`
}
