% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mean_na}
\alias{mean_na}
\title{Return the mean}
\usage{
mean_na(x)
}
\arguments{
\item{x}{data potentially with non-finite values}
}
\description{
Return the mean of `x`. If no observations, return `NA`. This is a workaround for the fact that in data.table, `:= mean()` will return `NaN` where there are no observations
}
