% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateutils.R
\name{month_days}
\alias{month_days}
\title{Number of days in a given month}
\usage{
month_days(year, month)
}
\arguments{
\item{year}{integer year value}

\item{month}{integer month value}
}
\description{
Get the number of days in a month given the year and month
}
