% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{month_to_date}
\alias{month_to_date}
\title{Get numeric month from string}
\usage{
month_to_date(x)
}
\arguments{
\item{x}{String vector of months, possibly including years}
}
\value{
If years are included, ebnd of month dates as.Date(). If years
        are omitted, numeric month-day values. Returns end of month date.
}
\description{
Get the numeric month value from string value months
}
\examples{
dates <- month_to_date(c("January 2014", "February 2014", "March 2014"))
}
