% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{number_finite}
\alias{number_finite}
\title{Number of finite values in a row}
\usage{
number_finite(Y)
}
\arguments{
\item{Y}{matrix like data object}
}
\description{
Return the number of finite values in a row of Y
}
