% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pct_chng}
\alias{pct_chng}
\title{Percent change}
\usage{
pct_chng(y, lag = 1)
}
\arguments{
\item{y}{data}

\item{lag}{number of periods for percent change}
}
\value{
The percentage change among the lag period
}
\description{
Calculate the percent change in `y` from one period to the next
}
\examples{
 
pct_chng(c(100,50,100,20,100,110))
}
