% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_utils.R
\name{pct_response}
\alias{pct_response}
\title{Percent of responses at a given frequency}
\usage{
pct_response(
  dt,
  col_name = NULL,
  by = c("month", "quarter", "week"),
  date_name = "ref_date"
)
}
\arguments{
\item{dt}{data table of responses}

\item{col_name}{name of column containing responses}

\item{by}{frequency of response aggregation, one of `"month"`, `"quarter"`, `"week"`}

\item{date_name}{name of column containing dates}
}
\description{
Return the percent of responses to categorical answers at a specified frequency
}
