% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateutils.R
\name{rollmax}
\alias{rollmax}
\title{Rolling Max}
\usage{
rollmax(x, n)
}
\arguments{
\item{x}{Numeric vector}

\item{n}{Integer span}
}
\value{
The maximum value of `x` with span `n`
}
\description{
Find the rolling maximum in `x` with span `n`
}
\examples{
rollmax(c(1,2,3), 2) ## c(2,3,3)
}
