% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rollmean}
\alias{rollmean}
\title{Rolling mean}
\usage{
rollmean(x, n)
}
\arguments{
\item{x}{data vector}

\item{n}{span of rolling mean}
}
\value{
Rolling mean of the input
}
\description{
Take the rolling mean of `x` over `n` elements
}
\examples{
rollmean(c(1,2,3),2) ## NA, 1.5, 2.5 
}
