% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateutils.R
\name{rollmin}
\alias{rollmin}
\title{Rolling Min}
\usage{
rollmin(x, n)
}
\arguments{
\item{x}{Numeric vector}

\item{n}{Integer span}
}
\value{
The minimum value of `x` with span `n`
}
\description{
Find the rolling minimum in `x` with span `n`
}
\examples{
rollmin(c(1,2,3),2) ## c(1,1,2)
}
