% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{row_to_list}
\alias{row_to_list}
\title{Convert rows to list}
\usage{
row_to_list(Y)
}
\arguments{
\item{Y}{matrix like data object}
}
\value{
Each row as a list
}
\description{
Return `Y` with each row as a list
}
\examples{
row_to_list(matrix(rnorm(20),10,2))
}
