% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sd_na}
\alias{sd_na}
\title{Return the standard deviation}
\usage{
sd_na(x)
}
\arguments{
\item{x}{data potentially with non-finite values}
}
\description{
Return the standard deviation of `x`. If no observations, return `NA`. This is a workaround for the fact that in data.table, `:= sd(x, na.rm = TRUE)` will return `NaN` where there are no observations
}
