% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seas_utils.R
\name{seas_df_wide}
\alias{seas_df_wide}
\title{Seasonally adjust long format data using seas()}
\usage{
seas_df_wide(df, sa_cols, x11 = FALSE, transfunc = "none")
}
\arguments{
\item{df}{wide format dataframe}

\item{sa_cols}{names or column indexes of series to seasonally adjust}

\item{x11}{T/F, use x11 as opposed to X-13 SEATS}

\item{transfunc}{Data transformation, one of `none` for no transformation, `auto` for automatic detection, or `log` for log transformation}
}
\description{
Seasonaly adjust multiple monthly or quarterly series in long format using X-13 SEATS via seas()
}
