% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seas_utils.R
\name{spline_fill}
\alias{spline_fill}
\title{Spline fill missing observations}
\usage{
spline_fill(x)
}
\arguments{
\item{x}{data with missing observations}
}
\value{
data with interpolated missing observations, except at head and tail, which remain NA
}
\description{
Spline fill missing observations from the first observation to the last, leaving NA observations in the head and tail
}
\examples{
spline_fill_trend(c(NA,1,2,3,NA,5)) ## NA 1 2 3 4 5
}
