% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seas_utils.R
\name{spline_fill_trend}
\alias{spline_fill_trend}
\title{Spline fill missing observations}
\usage{
spline_fill_trend(x)
}
\arguments{
\item{x}{data with missing observations}
}
\description{
Spline fill missing observations, designed for filling low frequency trend estimates
}
