% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sum_na}
\alias{sum_na}
\title{Return the sum}
\usage{
sum_na(x)
}
\arguments{
\item{x}{data potentially with non-finite values}
}
\description{
Return the sum of `x`. If no observations, return `NA`. This is a workaround for the fact that in data.table, `:= sum()` will return `NaN` where there are no observations
}
