% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seas_utils.R
\name{to_ts}
\alias{to_ts}
\title{Tabular data to ts() format}
\usage{
to_ts(x, dates)
}
\arguments{
\item{x}{data}

\item{dates}{dates}
}
\description{
transform data in 'x' corresponding to dates in 'dates' to ts() format
}
