% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seas_utils.R
\name{try_detrend}
\alias{try_detrend}
\title{Remove low frequency trends from data}
\usage{
try_detrend(x, outlier_rm = TRUE, span = 0.6)
}
\arguments{
\item{x}{data}

\item{outlier_rm}{T/F, remove outliers to estimate trends?}

\item{span}{span for the loess regression}
}
\value{
Data with trends removed
}
\description{
Estimate low frequency trends via loess regression and remove them. If the function errors, return x (i.e. no trend)
}
\examples{
try_detrend(c(1,3,6,7,9,11,14,15,17,18))
}
