% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seas_utils.R
\name{try_rm_outlier}
\alias{try_rm_outlier}
\title{Remove outliers}
\usage{
try_rm_outlier(x, z = 3, span = 0.6)
}
\arguments{
\item{x}{data}

\item{z}{cutoff value; outliers have absolute value greater than z*(s.d.)}

\item{span}{span for the loess regression to estimate trend}
}
\value{
data with outliers dropped
}
\description{
Identify outliers as deviations from trend and remove them
}
\examples{
try_rm_out(c(1,3,6,7,9,11,104,15,17,18))
}
