% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seas_utils.R
\name{can_seasonal}
\alias{can_seasonal}
\title{Can data be seasonally adjusted?}
\usage{
can_seasonal(dates)
}
\arguments{
\item{dates}{dates}
}
\description{
Return a logical indicating whether data at given dates can be seasonally adjusted using seas()
}
