% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_character}
\alias{extract_character}
\title{Extract character values}
\usage{
extract_character(x)
}
\arguments{
\item{x}{object containing character values}
}
\description{
Extract character values from x including space and underscore
}
