% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateutils.R
\name{fill_forward}
\alias{fill_forward}
\title{Fill Forward}
\usage{
fill_forward(x)
}
\arguments{
\item{x}{Transition matrix from a VAR model}
}
\value{
x with missing obs filled by forward value
}
\description{
Fill missing observations forward using the last finite observation
}
\examples{
fill_forward(c(1,2,NA,NA,3,NA,5)) ## 1 2 2 2 3 3 5   
}
