% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{first_of_month}
\alias{first_of_month}
\title{First of month}
\usage{
first_of_month(dates)
}
\arguments{
\item{dates}{A sequence of dates in `as.Date()` format}
}
\value{
First day of the month
}
\description{
Return the first day of the month for each date in `dates`
}
\examples{
dates <- seq.Date(from = as.Date("2020-09-11"),
                  by = "day", length.out = 10)
first_of_month(dates)
}
