% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateutils.R
\name{first_of_quarter}
\alias{first_of_quarter}
\title{First of Quarter}
\usage{
first_of_quarter(dates)
}
\arguments{
\item{dates}{Transition matrix from a VAR model}
}
\value{
The first day of the quarter for the dates
}
\description{
Find the first date in the quarter for a vector of dates
}
\examples{
first_of_quarter(as.Date("2019-9-15")) ## 2019-07-01
}
