% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateutils.R
\name{first_previous_quarter}
\alias{first_previous_quarter}
\title{First of previous quarter date}
\usage{
first_previous_quarter(date)
}
\arguments{
\item{date}{date value formated as.Date()}
}
\value{
The first day of the previous quarter of the date
}
\description{
Return the date of the first day of the previous quarter
}
\examples{
first_previous_quarter(as.Date("2019-09-15")) ## 2019-04-01
}
