% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_data_frq}
\alias{get_data_frq}
\title{Get frequency of data based on missing observations}
\usage{
get_data_frq(x, dates)
}
\arguments{
\item{x}{data, potentially with missing observations}

\item{dates}{corresponding dates in `as.Date()` format}
}
\description{
Guess the frequency of a data series based on the pattern of missing observations
}
