% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_in}
\alias{is_in}
\title{Find element of this_in that}
\usage{
is_in(that, this_in)
}
\arguments{
\item{that}{first object}

\item{this_in}{second object}
}
\description{
Find element of this_in that, ie `this_in%in%that`
}
