% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{last_in_week}
\alias{last_in_week}
\title{Last date in the week}
\usage{
last_in_week(dates)
}
\arguments{
\item{dates}{A sequence of dates in `as.Date()` format}
}
\value{
Last day of each week
}
\description{
Return the latest date in each week for the values in `dates`
}
\examples{
dates <- seq.Date(from = as.Date("2020-09-21"),
                  by = "day", length.out = 10)
last_in_week(dates)
}
