% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{limit_character}
\alias{limit_character}
\title{Limit Characters}
\usage{
limit_character(x, limit = 100)
}
\arguments{
\item{x}{object containing character values}

\item{limit}{maximum number of characters to return}
}
\value{
Character values within the limit
}
\description{
limit the number of characters in a string and remove spacial characters (will not drop numbers)
}
\examples{
limit_character("a\%b+&cd!efghij",limit = 3)  ## "abc"
}
