% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateutils.R
\name{long_run_var}
\alias{long_run_var}
\title{Long Run Variance of a VAR}
\usage{
long_run_var(A, Q, m, p)
}
\arguments{
\item{A}{Transition matrix from a VAR model in companion form}

\item{Q}{Covariance of shocks}

\item{m}{Number of series in the VAR}

\item{p}{Number of lags in the VAR}
}
\description{
Find the long run variance of a VAR using the transition equation `A` and shocks to observations `Q`
}
