% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seas_utils.R
\name{match_index}
\alias{match_index}
\title{Match index values}
\usage{
match_index(this, that)
}
\arguments{
\item{this}{first object}

\item{that}{second object}
}
\description{
Match index values of this to that
}
