% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateutils.R
\name{month_days}
\alias{month_days}
\title{Number of days in a given month}
\usage{
month_days(year, month)
}
\arguments{
\item{year}{integer year value}

\item{month}{integer month value}
}
\value{
The number of days in the month (integer)
}
\description{
Get the number of days in a month given the year and month
}
\examples{
month_days(2021,9) ## 30

month_days(2020,2) ## 29
}
