% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_utils.R
\name{total_response}
\alias{total_response}
\title{Number of of responses at a given frequency}
\usage{
total_response(dt, col_name = NULL, by = "month", date_name = "ref_date")
}
\arguments{
\item{dt}{data table of responses}

\item{col_name}{name of column containing responses}

\item{by}{frequency of response aggregation, one of `"month"`, `"quarter"`, `"week"`}

\item{date_name}{name of column containing dates}
}
\description{
Return the total number of responses to categorical answers at a specified frequency
}
