% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seas_utils.R
\name{try_sa}
\alias{try_sa}
\title{Seasonally adjust data using seas()}
\usage{
try_sa(x, dates, x11 = FALSE, transfunc = "none", series_name = NULL)
}
\arguments{
\item{x}{data}

\item{dates}{dates corresponding to data 'x'}

\item{x11}{T/F, use x11 as opposed to X-13 SEATS}

\item{transfunc}{Data transformation, one of `none` for no transformation, `auto` for automatic detection, or `log` for log transformation}

\item{series_name}{Include series name to print out if failure (for lapply() applications)}
}
\description{
Seasonaly adjust monthly or quarterly data using X-13 SEATS via seas()
}
