% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seas_utils.R
\name{try_trend}
\alias{try_trend}
\title{Estimate low frequnecy trends}
\usage{
try_trend(x, outlier_rm = TRUE, span = 0.6)
}
\arguments{
\item{x}{data}

\item{outlier_rm}{T/F, remove outliers to estimate trends?}

\item{span}{span for the loess regression}
}
\value{
Estimated trend in the data
}
\description{
Estimate low frequency trends via loess regression. If the function errors, return zeros (i.e. no trend)
}
\examples{
try_trend(c(1,3,6,7,9,11,14,15,17,18))
}
